requirement.txt
langchain-google-genai==4.1.2
langchain==1.2.0
python-dotenv==1.2.1
serpapi==0.1.5